// angular-biometric-designer/src/app/shared/models/helpdesk/cus.info.dto.ts
export class CusInfoDto {
    account_num: string = '';
    client_code: string = '';
    full_name: string = '';
    legal_id: string = '';
    birth_date: string = '';
    card_number: string = '';
    type_card: string = ''; // The Chinh/Phu
    date_active: string = '';
    status_card: string = ''; // Tinh trang the
    actual_account_balance: string = '';
    address_1: string = '';
    address_2: string = '';
    address_3: string = '';
    address_4: string = '';
    city: string = '';
    country: string = '';
    embossed_card_status: string = '';
    available_account_balance: string = '';
    product: string = '';
    date_expiry: string = '';
    name_print_card: string = '';
    branch: string = '';
    term_type: string = '';
}